sign-languages/yolo/deploy/code/requirements.txt
nvgpu
opencv-python
torchvision
seaborn
ultralytics
pynvml
omegaconf==2.3.0
Pillow